requirements.txt
Django==1.8.2
djangorestframework==3.1.2
Pillow==2.8.1
psycopg2==2.6
